fixed dynamic extrafield deletion in admin, colabo

+++ src/Controller/AdminController.php
@@ -11,51 +11,74 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class AdminController extends AbstractController
 {
+    /**
+     * Show the admin page
+     * 
+     * @param ManagerRegistry $doctrine
+     * 
+     * @return Response
+     */
     public function index(ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         $extrafieldsRepository = $doctrine->getRepository(ContactExtrafields::class);
         $extrafields = $extrafieldsRepository->findAll();
 
-        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         return $this->render('admin/index.html.twig', [
             'user' => $this->getUser(),
             'extrafields' => $extrafields,
         ]);
     }
 
+    /**
+     * Add, edit, and remove Extrafields
+     * 
+     * @param ManagerRegistry $doctrine
+     * @param Request $request
+     * 
+     * @return Response
+     */
     public function newExtrafields(ManagerRegistry $doctrine, Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         $extrafieldsRepository = $doctrine->getRepository(ContactExtrafields::class);
         $extrafields = $extrafieldsRepository->findAll();
         $entityManager = $doctrine->getManager();
 
         $post = $request->request;
         $postArray = $post->all();
+        $amountPost = count($postArray)/3; //we receive 3 fields (label,inputType,forCompany) for one extrafield, i could make it dynamic, but i don't have the time
+        $amountFields = count($extrafields);
 
-        $explodedKey = explode('_',array_keys($postArray)[count($postArray)-1]);
-        if(isset($explodedKey[1])){
-            $amount =  $explodedKey[1];
-        }else{
-            $this->addFlash('error', "Problèmes de découpage des extrafields");
-            return $this->render('admin/index.html.twig', [
-                'user' => $this->getUser(),
-                'extrafields' => $extrafields,
-            ]);
-        }
-
-        // dd($extrafields);
-        //dd($post->all());
-        for ($i=0; $i < $amount; $i++) { 
-            if(isset($postArray["label_". ($i+1)])){
-                $extrafield = isset($extrafields[$i]) ? $extrafields[$i] : $extrafield = new ContactExtrafields();
-                
-                $extrafield->setLabel($postArray["label_". ($i+1)]);
-                $extrafield->setInputType($postArray["inputType_". ($i+1)]);
-                $extrafield->setForCompany($postArray["forCompany_". ($i+1)] == "company" ? 1 : 0);
-    
-                isset($extrafields[$i]) ? null : $entityManager->persist($extrafield) ;
-            }else{
-                $entityManager->remove($extrafields[$i]);
+        if($amountPost > $amountFields){ //
+            for ($i=0; $i < $amountPost; $i++) { 
+                if(isset($postArray["label_". ($i+1)])){
+                    $extrafield = isset($extrafields[$i]) ? $extrafields[$i] : $extrafield = new ContactExtrafields();
+                    
+                    $extrafield->setLabel($postArray["label_". ($i+1)]);
+                    $extrafield->setInputType($postArray["inputType_". ($i+1)]);
+                    $extrafield->setForCompany($postArray["forCompany_". ($i+1)] == "company" ? 1 : 0);
+        
+                    isset($extrafields[$i]) ? null : $entityManager->persist($extrafield) ;
+                }else{
+                    if(isset($extrafields[$i])){
+                        $entityManager->remove($extrafields[$i]);
+                    }
+                }
+            }
+        }else{ // if we delete more element than we had at the start we have to slightly change some checks and counting
+            for ($i=0; $i < $amountFields; $i++) { 
+                if(isset($postArray["label_". ($i+1)])){
+                    $extrafield = isset($extrafields[$i]) ? $extrafields[$i] : $extrafield = new ContactExtrafields();
+                    
+                    $extrafield->setLabel($postArray["label_". ($i+1)]);
+                    $extrafield->setInputType($postArray["inputType_". ($i+1)]);
+                    $extrafield->setForCompany($postArray["forCompany_". ($i+1)] == "company" ? 1 : 0);
+        
+                    isset($extrafields[$i]) ? null : $entityManager->persist($extrafield) ;
+                }else{
+                    $entityManager->remove($extrafields[$i]);
+                }
             }
         }
